split SearchBar & SearchInput into separate files
